Size golf

+++ src/core/BalancerV3.sol
@@ -511,7 +511,7 @@ abstract contract BalancerV3 is SettlerAbstract, FreeMemory {
         */
 
         while (data.length >= _HOP_DATA_LENGTH) {
-            uint16 bps;
+            uint256 bps;
             assembly ("memory-safe") {
                 bps := shr(0xf0, calldataload(data.offset))
